# src/openapi.yaml
openapi: 3.0.3
info:
  title: Promiscope API
  description: >
    Backend API for Promiscope — a decentralised football scouting platform built on the Stellar
    blockchain (Soroban smart contracts).
  version: 1.0.0
  contact:
    name: Promiscope Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: API v1
  - url: /api/v2
    description: API v2 (includes SSE stream endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform
        operations covered by their granted scopes (#1019). Legacy keys without a scope list are
        unrestricted.
...
paths:
...
  /docs:
    get:
      tags:
        - docs
      summary: Returns the OpenAPI 3.0 specification as JSON.
      operationId: getDocs
      description: Returns the OpenAPI 3.0 specification as JSON.
      responses:
        '200':
          description: OpenAPI 3.0 document (application/json)
        '500':
          description: '{ success: false, error: string } - Spec file missing or unparsable'
      security: []
...
  /docs/ui:
    get:
      tags:
        - docs
      summary: Serves embedded Swagger UI.
      operationId: getDocsUi
      description: |-
        Serves embedded Swagger UI.
        Requires `swagger-ui-dist` to be installed:
          npm install swagger-ui-dist
      responses:
        '200':
          description: Swagger UI HTML page (text/html)
        '501':
          description: '{ success: false, error: string } - swagger-ui-dist not installed'
      security: []
  /docs/yaml:
    get:
      tags:
        - docs
      summary: Returns the raw OpenAPI YAML source.
      operationId: getDocsYaml
      description: Returns the raw OpenAPI YAML source.
      responses:
        '200':
          description: Raw OpenAPI YAML (text/yaml)
        '500':
          description: '{ success: false, error: string } - Spec file missing'
      security: []
...